% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{eco_lm}
\alias{eco_lm}
\title{Estimating Linear Regressions}
\usage{
eco_lm(data, formula)
}
\arguments{
\item{data}{A data.frame or tibble}

\item{formula}{An object of class \code{\link[stats]{formula}}}
}
\description{
Hello
}
