% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econometrics2-package.R
\docType{package}
\name{econometrics2-package}
\alias{econometrics2}
\alias{econometrics2-package}
\title{econometrics2: Provides all neccessary tools for the courses "Probability theory and statistics", "Econometrics I" and
"Econometrics II" at University of Copenhagen}
\description{
The packages provides all neccessary tools to answer questions in the courses mentioned above at
University of Copenhagen. The package aim at letting upcomming students use R instead or as supplement to
Stata and Oxmetrics.
}
\author{
\strong{Maintainer}: Jakob Jul Elben \email{elbenjakobjul@gmail.com}

Authors:
\itemize{
  \item Rasmus Salmon \email{rasmussalmon@gmail.com}
}

}
